Refactor(SE-75): Refactor a few items and clean up
