+++ src/main/java/com/finalproject/stayease/config/SecurityConfig.java
@@ -71,9 +71,10 @@ public class SecurityConfig {
   private void configureAuthorization(AuthorizeHttpRequestsConfigurer<HttpSecurity>.AuthorizationManagerRequestMatcherRegistry auth) {
     auth.requestMatchers("/api/v1/role").denyAll();
     auth.requestMatchers(HttpMethod.GET, "/api/v1/properties/", "/api/v1/properties/rooms/"). permitAll();
+    auth.requestMatchers(HttpMethod.PUT, "/api/v1/users/profile/email").permitAll();
     auth.requestMatchers("/api/v1/role/user").hasRole("USER");
     auth.requestMatchers("/api/v1/role/tenant", "/api/v1/properties/**").hasRole("TENANT");
-    auth.requestMatchers("/api/v1/auth/**", "/login/**", "/oauth2/**").permitAll();
+    auth.requestMatchers("/api/v1/auth/**", "api/v1/auth/login/**", "api/v1/oauth2/**", "/register**").permitAll();
     auth.anyRequest().authenticated();
   }
